"first commit with hamzaafterchat"

+++ src/main/java/fullStack/template/repository/UserAppRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface UserAppRepo extends JpaRepository<UserApp,Integer> {
+public interface UserAppRepo extends JpaRepository<UserApp,Long> {
     public UserApp findByEmail(String email);
 }
